Start splitting different command options into separate files
